+++ src-tauri/src/live/opcodes_models.rs
@@ -47,12 +47,12 @@ pub struct Entity {
     pub hits_heal: i64,
     pub skill_uid_to_heal_skill: HashMap<i32, Skill>,
     // Boss Damage
-    pub total_dmg_boss: u128,
-    pub crit_total_dmg_boss: u128,
-    pub crit_hits_dmg_boss: u128,
-    pub lucky_total_dmg_boss: u128,
-    pub lucky_hits_dmg_boss: u128,
-    pub hits_dmg_boss: u128,
+    pub total_dmg_boss: i64,
+    pub crit_total_dmg_boss: i64,
+    pub crit_hits_dmg_boss: i64,
+    pub lucky_total_dmg_boss: i64,
+    pub lucky_hits_dmg_boss: i64,
+    pub hits_dmg_boss: i64,
     pub skill_uid_to_dmg_skill_boss: HashMap<i32, Skill>,
 }
